pyproject.toml
[project]
name = "innovative-bigtech-cryptography-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.122.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]

[tool.ruff]
# Установка длины строки
line-length = 79

# Исключение файлов и папок
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Целевая версия Python
target-version = "py313"

[tool.ruff.lint]
# Включение групп правил — https://docs.astral.sh/ruff/rules/
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "B",   # flake8-bugbear

    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "RUF", # ruff-specific rules
    "SIM", # lake8-simplify
]


# Игнорирование определенных правил
ignore = [
    "RUF001",  # кириллица вместо англ символов
    "RUF003",  # кириллица вместо англ символов
]

# Разрешение автоисправления для определенных правил
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]

# Использование двойных кавычек
quote-style = "double"

# Отступы в 4 пробела (по умолчанию)
indent-style = "space"

# Сохранение магических запятых
skip-magic-trailing-comma = false
